requirements.dev.txt
asgiref==3.4.1
astroid==2.8.0
Django==3.2.7
django-extensions==3.1.3
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.13.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
pep8==1.7.1
platformdirs==2.3.0
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.2
wrapt==1.12.1
